+++ packages/happn-cluster/RELEASES.md
@@ -330,5 +330,4 @@ non-release 2017-12-14
 
 12.2.4 2022-11-26
 -----------------
-  - TEN-4: productionize summon (authType saved to user)
-  - TEN-140: outdated pem module
+  - Sqlite db provider